Object
This code demonstrates inheritance, polymorphism, and encapsulation. The Student class inherits the methods and attributes of the Person class. Polymorphism is shown using object1, which is class Student, by printing out getFirstName() which is possible because obejct1 still has the methods of its parent class, Person. Encapsulation is shown by changing the id of object1 thorugh a setter rather than accessing it directly.